+++ MiniTwitch.Helix/HelixWrapper.cs
@@ -907,4 +907,8 @@ public class HelixWrapper
         CancellationToken cancellationToken = default)
     => _all.GetGlobalChatBadges(cancellationToken);
 
+    public Task<HelixResult<ModeratedChannels>> GetModeratedChannels(
+        int? first = null,
+        CancellationToken cancellationToken = default)
+    => _all.GetModeratedChannels(first, cancellationToken);
 }