+++ MiniTwitch.Helix/Models/ModerationCategory.cs
@@ -114,4 +114,9 @@ public sealed class ModerationCategory
         long broadcasterId,
         CancellationToken cancellationToken = default)
     => _all.GetShieldModeStatus(broadcasterId, cancellationToken);
+
+    public Task<HelixResult<ModeratedChannels>> GetModeratedChannels(
+        int? first = null,
+        CancellationToken cancellationToken = default)
+    => _all.GetModeratedChannels(first, cancellationToken);
 }